✨ Feat: 카테고리에 저장한 장소 조회 API 연동
